Rename core.py to _core.py, enum.py to _enum.py
