+++ client/.gitignore
@@ -1,5 +1,5 @@
 ### Generated files ###
-enums.py
+_enums.py
 
 ### Poetry ###
 .poetry
