+++ client/h2ogpt_client/__init__.py
@@ -1,4 +1,4 @@
-from h2ogpt_client.core import Client
-from h2ogpt_client.enums import LangChainMode, PromptType
+from h2ogpt_client._core import Client
+from h2ogpt_client._enums import LangChainMode, PromptType
 
 __all__ = ["Client", "PromptType", "LangChainMode"]
