+++ client/h2ogpt_client/_core.py
@@ -4,7 +4,7 @@ from typing import Any, Dict, List, Optional, OrderedDict, Tuple, ValuesView
 
 import gradio_client  # type: ignore
 
-from h2ogpt_client import enums
+from h2ogpt_client import _enums
 
 
 class Client:
@@ -49,7 +49,7 @@ class TextCompletionCreator:
 
     def create(
         self,
-        prompt_type: enums.PromptType = enums.PromptType.plain,
+        prompt_type: _enums.PromptType = _enums.PromptType.plain,
         input_context_for_instruction: str = "",
         enable_sampler=False,
         temperature: float = 1.0,
@@ -63,7 +63,7 @@ class TextCompletionCreator:
         repetition_penalty: float = 1.07,
         number_returns: int = 1,
         system_pre_context: str = "",
-        langchain_mode: enums.LangChainMode = enums.LangChainMode.DISABLED,
+        langchain_mode: _enums.LangChainMode = _enums.LangChainMode.DISABLED,
     ) -> "TextCompletion":
         """
         Creates a new text completion.
@@ -165,7 +165,7 @@ class ChatCompletionCreator:
 
     def create(
         self,
-        prompt_type: enums.PromptType = enums.PromptType.plain,
+        prompt_type: _enums.PromptType = _enums.PromptType.plain,
         input_context_for_instruction: str = "",
         enable_sampler=False,
         temperature: float = 1.0,
@@ -179,7 +179,7 @@ class ChatCompletionCreator:
         repetition_penalty: float = 1.07,
         number_returns: int = 1,
         system_pre_context: str = "",
-        langchain_mode: enums.LangChainMode = enums.LangChainMode.DISABLED,
+        langchain_mode: _enums.LangChainMode = _enums.LangChainMode.DISABLED,
     ) -> "ChatCompletion":
         """
         Creates a new chat completion.
